requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
pytest-django==4.8.0

pillow==10.2.0